refactor(dashboard): move statusColor to format.ts alongside statusIcon
statusColor and statusIcon are logically paired mapping functions.
Co-locating them in format.ts makes the module the single source of
truth for status display mappings and removes the export from the
React component file.
